+++ R/highlevel64.R
@@ -15,7 +15,7 @@
 #' @param timefun a function for timing such as [bit::repeat.time()] or [system.time()]
 #' @param what a vector of names of high-level functions
 #' @param uniorder one of the order parameters that are allowed in [unique.integer64()] and [unipos.integer64()]
-#' @param taborder one of the order parameters that are allowed in [table.integer64()]
+#' @param taborder one of the order parameters that are allowed in [table()]
 #' @param plot set to FALSE to suppress plotting
 #'
 #' @details
@@ -186,7 +186,7 @@ benchmark64 <- function(nsmall=2L^16L, nbig=2L^25L, timefun=repeat.time) {
   })[3L]
   message('since not unique, then check distribution of frequencies')
   tim1[i] <- tim1[i] + timefun({
-   if (i==1L) tabulate(table(b, exclude=NULL)) else tabulate(table.integer64(b, return='list')$counts)
+   if (i==1L) tabulate(table(b, exclude=NULL)) else tabulate(table(b, return='list')$counts)
   })[3L]
   message("OK, let's plot the percentiles of unique values versus the percentiles allowing for duplicates")
   tim1[i] <- tim1[i] + timefun({
@@ -313,7 +313,7 @@ benchmark64 <- function(nsmall=2L^16L, nbig=2L^25L, timefun=repeat.time) {
 
   message(colnames(tim2)[i], " table(b)")
   tim2["table(b)", i] <- timefun({
-   if (i==1L) table(b) else table.integer64(b, return='list')
+   if (i==1L) table(b) else table(b, return='list')
   })[3L]
 
   message(colnames(tim2)[i], " sort(b)")
@@ -996,14 +996,12 @@ optimizer64 <- function(nsmall=2L^16L,
    tim["table", "both"] <- timefun({
     p <- table(x, exclude=NULL)
    })[3L]
-   p <- p[-length(p)]
 
    x <- as.integer64(x)
 
    tim["table.64", "both"] <- timefun({
-    p2 <- table.integer64(x, order=taborder)
+    p2 <- table(x, exclude=NULL, order=taborder)
    })[3L]
-   p2 <- p2[-1L]
    stopifnot(identical(p2, p))
 
    tim["hashmaptab", "both"] <- timefun({
@@ -1076,27 +1074,27 @@ optimizer64 <- function(nsmall=2L^16L,
 
             hashcache(x)
             tim["hash.cache", "use"] <- timefun({
-                p <- table.integer64(x, order=taborder)
+                p <- table(x, exclude=NULL, order=taborder)
             })[3L]
             remcache(x)
 
-            sortordercache(x)
+            sortordercache(x, na.last=TRUE)
             tim["sort.cache", "use"] <- timefun({
-                p2 <- table.integer64(x, order=taborder)
+                p2 <- table(x, exclude=NULL, order=taborder)
             })[3L]
             stopifnot(identical(p2, p))
             remcache(x)
 
-            ordercache(x)
+            ordercache(x, na.last=TRUE)
             tim["order.cache", "use"] <- timefun({
-                p2 <- table.integer64(x, order=taborder)
+                p2 <- table(x, exclude=NULL, order=taborder)
             })[3L]
             stopifnot(identical(p2, p))
             remcache(x)
 
             if (plot) {
                 barplot(t(tim), cex.names=0.7)
-                title(paste0("table.integer64(", n, ", order=", taborder, ")"))
+                title(paste0("table(", n, ", order=", taborder, ")"))
             }
 
             ret[["table", as.character(n)]] <- tim
@@ -1674,7 +1672,7 @@ duplicated.integer64 <- function(x, incomparables = FALSE, nunique = NULL, metho
 #'   the result has no names).
 #'
 #' @seealso [unique()] for the generic, [unipos()] which gives the indices
-#'   of the unique elements and [table.integer64()] which gives frequencies
+#'   of the unique elements and [table()] which gives frequencies
 #'   of the unique elements.
 #'
 #' @examples
@@ -1955,23 +1953,21 @@ unipos.integer64 <- function(x,
 #'
 #' `table.integer64` uses the cross-classifying integer64 vectors to build a
 #'   contingency table of the counts at each combination of vector values.
-#'
+#'   
 #' @param ... one or more objects which can be interpreted as factors
 #'   (including character strings), or a list (or data frame) whose
 #'   components can be so interpreted.  (For `as.table` and `as.data.frame`,
 #'   arguments passed to specific methods.)
-#' @param nunique NULL or the number of unique values of table (including NA).
-#'   Providing `nunique` can speed-up matching when `table` has no cache. Note
-#'   that a wrong `nunique` can cause undefined behaviour up to a crash.
-#' @param order By default results are created sorted by "values", or by "counts"
-#' @param method NULL for automatic method selection or a suitable low-level
-#'   method, see details
-#' @param return choose the return format, see details
+#' @param exclude levels to remove for all factors in ....
+#' @param useNA whether to include NA values in the table.
 #' @param dnn the names to be given to the dimensions in the result
 #'   (the _dimnames names_).
 #' @param deparse.level controls how the default `dnn` is constructed. See Details.
 #'
 #' @details
+#' If one argument of ... is `integer64` the `table.integer64` method is used. Only this 
+#' method supports the arguments `return`, `order`, `nunique`, and `method`.
+#' 
 #' This function automatically chooses from several low-level functions considering
 #'   the size of `x` and the availability of a cache.
 #'
@@ -1987,9 +1983,6 @@ unipos.integer64 <- function(x,
 #'   remaining arguments, `deparse.level = 0` gives an empty name,
 #'   `deparse.level = 1` uses the supplied argument if it is a symbol,
 #'   and `deparse.level = 2` will deparse the argument.
-#'
-#' Arguments `exclude`, `useNA`, are not supported, i.e. `NA`s are always tabulated,
-#'   and, different from [table()] they are sorted first if `order="values"`.
 #'
 #' @return By default (with `return="table"`) [table()] returns a
 #'   _contingency table_, an object of class `"table"`, an array of integer values.
@@ -2035,26 +2028,86 @@ unipos.integer64 <- function(x,
 #' x <- as.integer64(sample(c(rep(NA, 9), 1:9), 32, TRUE))
 #' y <- as.integer64(sample(c(rep(NA, 9), 1:9), 32, TRUE))
 #' z <- sample(c(rep(NA, 9), letters), 32, TRUE)
-#' table.integer64(x)
-#' table.integer64(x, order="counts")
-#' table.integer64(x, y)
-#' table.integer64(x, y, return="data.frame")
+#' table(x)
+#' table(x, order="counts")
+#' table(x, y)
+#' table(x, y, return="data.frame")
 #'
 #' message("via as.integer64.factor we can use 'table.integer64' also for factors")
-#' table.integer64(x, as.integer64(as.factor(z)))
+#' table(x, as.integer64(as.factor(z)))
 #' @keywords category
 #' @concept counts
 #' @concept frequencies
 #' @concept occurrences
 #' @concept contingency table
 #' @export
-table.integer64 <- function(...,
+table = function(..., exclude=if (useNA == "no") c(NA, NaN), useNA=c("no", "ifany", "always"), dnn=list.names(...), deparse.level=1L) UseMethod("table")
+#' @exportS3Method table default
+table.default = function(..., exclude=if (useNA == "no") c(NA, NaN), useNA=c("no", "ifany", "always"), dnn=list.names(...), deparse.level=1L) {
+  if (...length() && any(unlist(lapply(list(...), is.integer64)))) {
+    withCallingHandlers({
+      sys_call = sys.call()
+      sys_call[[1L]] = table.integer64
+      eval(sys_call, envir = parent.frame())},
+      error = function(e) {stop(errorCondition(e$message, call=choose_sys_call(c("table", "table.default"), "table.integer64")))}, 
+      warning = function(w) {
+        warning(warningCondition(w$message, call=choose_sys_call(c("table", "table.default"), "table.integer64")))
+        invokeRestart("muffleWarning")
+      }
+    )
+  } else {
+    withCallingHandlers({
+      sys_call = sys.call()
+      sys_call[[1L]] = base::table
+      eval(sys_call, envir = parent.frame())},
+      error = function(e) {stop(errorCondition(e$message, call=choose_sys_call(c("table", "table.default"))))}, 
+      warning = function(w) {
+        warning(warningCondition(w$message, call=choose_sys_call(c("table", "table.default"))))
+        invokeRestart("muffleWarning")
+      }
+    )
+  }
+}
+choose_sys_call = function(function_names, name_to_display=NULL) {
+  sc = sys.calls()
+  sc_length = length(sc)
+  if (sc_length == 1L || length(function_names) == 0L) return(sc[[1L]])
+  sc_char = gsub("^(.*::)?(.+)[(].*", "\\2", as.character(sc))
+  sel = rev(which(sc_char == function_names[length(function_names)]))[1L]
+  if (is.na(sel)) 
+    sel = 1L
+  for (i in rev(seq_along(function_names))[-1L]) {
+    if (sel == 1L || sc_char[sel - 1L] != function_names[i])
+      break
+    sel = sel - 1L
+  }
+  sc = sc[[sel]]
+  if (!is.null(name_to_display))
+    sc[[1L]] = as.name(name_to_display)
+  sc
+}
+
+#' @method table integer64
+#' @param return choose the return format, see details
+#' @param order By default results are created sorted by "values", or by "counts"
+#' @param nunique NULL or the number of unique values of table (including NA).
+#'   Providing `nunique` can speed-up matching when `table` has no cache. Note
+#'   that a wrong `nunique` can cause undefined behaviour up to a crash.
+#' @param method NULL for automatic method selection or a suitable low-level
+#'   method, see details
+#' @aliases table.integer64
+#' @exportS3Method table integer64
+#' @rdname table
+table.integer64 = function(...,
                             return = c("table", "data.frame", "list"),
                             order = c("values", "counts"),
                             nunique = NULL,
                             method = NULL,
+                            exclude = if (useNA == "no") c(NA, NaN),
+                            useNA = c("no", "ifany", "always"),
                             dnn = list.names(...),
                             deparse.level = 1L) {
+
   order <- match.arg(order)
   return <- match.arg(return)
   # this is taken from 'table'
@@ -2085,7 +2138,7 @@ table.integer64 <- function(...,
   A <- function(i) eval(argsymbols[[i]], argframe)
   N <- length(argsymbols)
   if (!N)
-      stop("nothing to tabulate")
+      stop("nothing to tabulate", domain="R-base")
   if (N == 1L && is.list(A(1L))) {
     args <- A(1L) # nolint: object_overwrite_linter. This code should probably be refactored anyway.
     if (length(dnn) != length(args))
@@ -2109,10 +2162,10 @@ table.integer64 <- function(...,
     d <- integer64(N+1L); d[[1L]] <- 1L
     dims <- vector("list", N)
     names(dims) <- dnn
-    for (i in 1:N) {
+    for (i in seq_len(N)) {
       a <- A(i)
       if (length(a) != n)
-        stop("all input vectors must have the same length")
+        stop("all arguments must have the same length", domain="R-base")
       if (!is.integer64(a)) {
         warning("coercing argument ", i, " to integer64")
         a <- as.integer64(a)
@@ -2121,7 +2174,7 @@ table.integer64 <- function(...,
       if (is.null(cache_env$order)) {
         s <- clone(a)
         o <- seq_along(s)
-        ramsortorder(s, o)
+        ramsortorder(s, o, na.last=TRUE)
         nu[[i]] <- sortnut(s)[["nunique"]]
       } else if (is.null(cache_env$sort)) {
         o <- cache_env$order
@@ -2196,7 +2249,7 @@ table.integer64 <- function(...,
     sorttab={
       if (is.null(cache_env) || is.null(cache_env$sort)) {
         s <- clone(x)
-        ramsort(s, na.last=FALSE)
+        ramsort(s, na.last=TRUE)
       } else {
         s <- get("sort", cache_env, inherits=FALSE)
       }
@@ -2208,7 +2261,7 @@ table.integer64 <- function(...,
     ordertab={
       if (is.null(cache_env) || is.null(cache_env$order)) {
         o <- seq_along(x)
-        ramorder(x, o, na.last=FALSE)
+        ramorder(x, o, na.last=TRUE)
       } else {
         o <- get("order", cache_env, inherits=FALSE)
       }
@@ -2222,7 +2275,7 @@ table.integer64 <- function(...,
   if (order=="values") {
     if (startsWith(method, "hash")) {
       o <- seq_along(val)
-      ramsortorder(val, o, na.last=FALSE)
+      ramsortorder(val, o, na.last=TRUE)
       cnt <- cnt[o]
     }
   } else {
@@ -2246,6 +2299,17 @@ table.integer64 <- function(...,
         a <- array(0L, dim=nu, dimnames=lapply(dims, as.character))
         a[as.integer(val)+1L] <- as.integer(cnt)
         cnt <- a
+      }
+      useNA = match.arg(useNA, c("no", "ifany", "always"))
+      if (length(exclude) > 0L) {
+        sel = suppressWarnings(lapply(dimnames(cnt), Negate(`%in%`), table=exclude))
+        cnt = do.call(`[`, c(list(x=cnt), sel, list(drop=FALSE)))
+      }
+      if (useNA == "always") {
+        dimnames_new = lapply(dimnames(cnt), function(el) if(any(is.na(el))) el else c(el, NA))
+        cnt_new = array(0L, as.integer(vapply(dimnames_new, length, 0L)))
+        dimnames(cnt_new) = dimnames_new
+        cnt = do.call(`[<-`, c(list(x=cnt_new), lapply(dim(cnt), seq_len), list(value=cnt)))
       }
       oldClass(cnt) <- "table"
     },